+++ CHANGES.rst
@@ -23,6 +23,8 @@ Unreleased
     switching ``POST`` to ``GET``. This preserves the current behavior of
     ``GET`` and ``POST`` redirects, and is also correct for frontend libraries
     such as HTMX. :issue:`5895`
+-   ``provide_automatic_options=True`` can be used to enable it for a view when
+    it's disabled in config. Previously, only disabling worked. :issue:`5916`
 
 
 Version 3.1.2